+++ gems/aws-sdk-apigateway/lib/aws-sdk-apigateway/client.rb
@@ -7061,7 +7061,7 @@ module Aws::APIGateway
         params: params,
         config: config)
       context[:gem_name] = 'aws-sdk-apigateway'
-      context[:gem_version] = '1.92.0'
+      context[:gem_version] = '1.93.0'
       Seahorse::Client::Request.new(handlers, context)
     end